+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - add bool `callCubeDispose` in `CubeBuilder` `to decide call `dispose` or not in cube.
 - add method `listen` inner `Cube` to listen `ObservableValue`.
+- add method `listenActions` inner `Cube` to listen `Actions`.
 
 # 0.6.0